remove non-story
If you listen to the actual audio at the link provided, I really don't think it sounds like Trump is actually inviting them to sit in on the interviews... It really just sounds like he will be doing the interviews privately, but that the club will still be open so people should come by to golf and laugh at how much press is there. A bunch of outlets reported this as a story, but I just don't see how they can make the claims that he invited them to sit in on interviews. So proposing to remove this story unless others disagree or have more evidence.