Fix improper use of 'while'
'While' should only be used to start a sentence in the form "While [some clause], [some related but different clause]". I think what you were going for was the other common usage of "[Some clause], while [some related but different clause]". That could be used here, but it'd make for quite the long sentence and impede readability.