Update 2017-11-02-day-287.markdown
Check #1. As I read the NYT story, the mortgage interest deduction is capped at $500K for new mortgages (down from $1 million), but it's a little confusing because the deduction is for interest payments while the limit is for the whole mortgage. And good catch -- "repeals the state and local income taxes" was fixed before I got there. :-)