Update 2017-11-10-day-295.markdown
I hope I didn't mangle #6, but "never tried a case for a lifetime appointment" didn't quite work for me. I'm also sort of in the mood to stick it to Alabama.

I'm also not 100% sure of #2. It sounds as though technically the GOP senatorial committee withdrew from the joint committee, yes?