Update 2017-12-18-day-333.markdown
On #3, I suggest putting the whole phrase "my people" in quotes because quotes just around "people" look like scare quotes, suggesting e.g. that "people" is a euphemism. (Which it may be, but I'll give them the benefit of the doubt. For now.)

On Notables #1 -- In the Politico story, I only count one senator as urging Franken to reconsider: Manchin. Leahy and unnamed have expressed regrets about calling for Franken's resignation, in private or off the record, but don't seem to have called for him to reconsider. Neither has, apparently, Lankford,  "At least four senators express reservations about Franken's decision to resign" is a possible alternative.