Update 2018-01-10-day-356.markdown
Re #2: Do we know what state's court is referred to here? I didn't see it in either the ABC or CNN story (sloppy, sloppy!). Probably wherever BuzzFeed is based but IDKWTF that is.

In #6, it only "reverses" the decision with regard to Florida. Other wordings are possible, e.g., "limits the Trump administration's plan . . ."

The fourth Notable cracks me up. Sense of order? Good luck with that.