Update 2018-01-11-day-357.markdown
#1: I changed "will" to "may". I think the distinction is important.

Re #3: Do we believe DJT is grammatically capable of beginning a tweet with "With that being said"? ;-)

#4: According to the source story, there WAS discussion of this but the idea was "scuttled."

Notables 2: Suggest deleting the last sentence because it seems only Blankenship called himself a "political prisoner" and in any case it doesn't seem that any of them are currently being "held," i.e., in jail.