Update 2018-03-02-day-407.markdown
Notable #1 -- I think it's important to get the Utah connection in there, OK?

Notable #3 -- The bit about the Russian mafia threat isn't clear in the CNN story either. Russians living in the Boston area might be angry if Mueller targets them? Huh?

Today gets at least an 8.5 out of 10 on the WTF scale. Maybe even a 9.