Update 2018-03-13-day-418.markdown
WTF! The power goes out for 3 hrs 15 mins and when it comes back on Tillerson's been fired, an ICE spokesman has resigned, and 45's "personal assistant" has been canned -- or maybe being seconded to the re-election committee is a promotion? (We're having our third major nor'easter in two weeks. This one comes with snow.)