Update 2018-03-20-day-425.markdown
#2 -- WaPo story now says that according to a top exec at this firm Olson is not considering the offer. This may have been superseded? Story doesn't refer to a previous offer so maybe this is it.

Notable #3 -- "in which their platform WAS MISUSED in the run-up to the 2018 midterm elections"? Past tense is odd because the run-up is just getting under way. Unfortunately it's screwy in the source too. :-(

Notable #4 -- "contractors" doesn't sound as exciting as "sex workers." ;-) Placement of "on behalf of political clients" looked odd so I suggest deleting it.