Update 2018-03-29-day-434.markdown
In the California judge item that follows #3, I broke up the sentence and put the company's name in because the clause put the verb an awful long way from the subject.

Notable #2: I may have muddied the waters here, but according to the source the FBI was brought in by the Latvian govt and not as a direct result of Trump & Ivanka's meeting with Krutoy. (It's not clear that meeting was in 2010 but it could have been.)

Notable #4: I do hope they invite the bunny.