Update 2018-05-14-day-480.markdown
#1 -- Even Chicago says it's OK to change case in a quote to fit the syntax. :-)

#5 -- I'm pretty sure that "his" refers to Pence, but please check.

In Notable #7, I'm not sure "penalized" is the best word, but "after US suppliers were blocked from selling parts to the Chinese company because it violated US sanctions against North Korea and Iran" might be TMI. I'm also not sure what "export privileges" refers to -- it was US suppliers that were exporting to China. ZTE seems to have been exporting to Iran and NKorea, but that was a no-no.