Update day-614.markdown
Sorry about the last one - 
I'm assuming you let me know the last one was inappropriate because I was getting too into it on the last one and I shouldn't reference the people who are not up for the supreme court nomination. I still think its really interesting to see a source which has the actual yearbook page which lets the reader actually get some sort of a glimpse into Kavanaugh's character in highschool, which is interesting because it seems to fly directly in the face of his fox news interview where he claims he spent most of his time studying.