Update 2018-09-27-day-616.markdown
#5 -- that comma (when Brett Kavanaugh, shoved her friend . . .) is actually in the original. It's probably safe to take it out.

#7 -- The source now says four, so I added the fourth, Phil Scott of VT.

Notable #2 -- I think it's best to end with "at the Justice Department." IIRC this has at least something to do with the NYT story about wearing a wire and invoking the 25th Amendment, not just whether RR was going to resign or get fired.

That POTUS presser was more dada than dada. Wow.