Update 2018-10-02-day-621.markdown
On Notable #5, as I read it, the source says that these ethics complaints deal with events before the hearing, but it also says that he's likely to face additional questions as a result of his behavior at the Sept. 27 hearing. I rewrote that bit -- see if it still makes sense.